Separate API from app.py and clean the code
